pyproject.toml
[tool.poetry]
name = "sauce"
version = "0.1.0"
description = ""
authors = ["Skogen <muumlover@live.com>"]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^21.9.3"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.17.8"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
